requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langsmith>=0.0.6
openai>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv
uvicorn
fastapi
streamlit
click
pytest
tqdm
unstructured>=0.14.0
pdfminer.six>=20221105
pi_heif
unstructured_inference
pdf2image
pytesseract
unstructured_pytesseract
